Inclusao de conexao de BD na databse.js e correçoes de digitaçao
