+++ src/data/database.js
@@ -1,0 +1,11 @@
+//conectar o banco de dados 
+async function connect(){
+    if(global.connection && global.connection.state !== 'disconnected')
+        return global.connection;
+
+    const mysql = require("mysql2/promise");
+    const connection = await mysql.createConnection("mysql://root:luiztools@localhost:3306/crud");
+    console.log("Conectou no MySQL!");
+    global.connection = connection;
+    return connection;
+}
