+++ src/routes/routes.js
@@ -10,6 +10,6 @@ routes.delete('/removeProduct/:user_id')//Remover produto da lista de clientes
 routes.delete('/listProduct')//Consultar todos os produtos disponiveis
 
 routes.post('/checkout/:user_id')//Finalizar compras
-routes.post('/uerHistory/:user_id')//Consultar a lista de compras do cliente
+routes.post('/userHistory/:user_id')//Consultar a lista de compras do cliente
 
 routes.get('/listShop')//Listar lojas
